+++ packages/osd-monaco/src/json/worker/json.worker.ts
@@ -28,35 +28,6 @@
  * under the License.
  */
 
-const path = require('path');
-
-const createLangWorkerConfig = (lang) => ({
-  mode: 'production',
-  entry: path.resolve(__dirname, 'src', lang, 'worker', `${lang}.worker.ts`),
-  output: {
-    path: path.resolve(__dirname, 'target/public'),
-    filename: `${lang}.editor.worker.js`,
-  },
-  resolve: {
-    modules: ['node_modules'],
-    extensions: ['.js', '.ts', '.tsx'],
-  },
-  stats: 'errors-only',
-  module: {
-    rules: [
-      {
-        test: /\.(js|ts)$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            babelrc: false,
-            presets: [require.resolve('@osd/babel-preset/webpack_preset')],
-          },
-        },
-      },
-    ],
-  },
-});
-
-module.exports = [createLangWorkerConfig('xjson'), createLangWorkerConfig('json')];
+/* eslint-disable */
+// @ts-ignore
+import 'monaco-editor/esm/vs/language/json/json.worker.js';